[SPARK-41102][CONNECT][REFACTORING] Merge SparkConnectPlanner and SparkConnectCommandPlanner.
